+++ src/test/php/lang/partial/unittest/Isbn.class.php
@@ -1,7 +1,9 @@
 <?php namespace lang\partial\unittest;
 
-use lang\partial\ValueObject;
-use lang\partial\WithCreation;
+use lang\partial\Accessors;
+use lang\partial\ToString;
+use lang\partial\Equals;
+use lang\partial\Builder;
 
 /**
  * Used by InstanceCreationTest
@@ -9,8 +11,10 @@ use lang\partial\WithCreation;
 class Isbn extends \lang\Object {
   const EAN13 = 13;
 
-  use Isbn\including\ValueObject;
-  use Isbn\including\WithCreation;
+  use Isbn\including\Accessors;
+  use Isbn\including\ToString;
+  use Isbn\including\Equals;
+  use Isbn\including\Builder;
 
   private $number, $type;