+++ src/test/php/lang/partial/unittest/Walls.class.php
@@ -6,5 +6,5 @@ use lang\partial\ListOf;
  * Used as fixture in the "ListOfTest" class
  */
 class Walls extends \lang\Object implements \IteratorAggregate {
-  use ListOf;
+  use Walls\is\ListOf;
 }